+++ src/StatisticsFunctions/kernels/plp_var_q16s_xpulpv2.c
@@ -73,9 +73,8 @@ void plp_var_q16s_xpulpv2(
                          const int16_t * __restrict__ pSrc,
                          uint32_t blockSize,
                          uint32_t deciPoint,
-                         int32_t * __restrict__ pRes){
+                         int16_t * __restrict__ pRes){
     
- 
   int32_t square_of_mean;
   int32_t square_of_values;
